requirements.txt
# Add additional python dependencies here
pyyaml
mysqlclient
coverage